+++ roxctl/common/flags.go
@@ -50,8 +50,10 @@ func GetGRPCConnection() (*grpc.ClientConn, error) {
 
 	if !flags.UseDirectGRPC() {
 		opts.DialTLS = func(ctx context.Context, endpoint string, tlsClientConf *tls.Config, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
-			return http1DowngradeClient.ConnectViaProxy(ctx, endpoint, tlsClientConf, http1DowngradeClient.ExtraH2ALPNs(alpn.PureGRPCALPNString), http1DowngradeClient.DialOpts(opts...))
+			return http1DowngradeClient.ConnectViaProxy(ctx, endpoint, tlsClientConf, http1DowngradeClient.ForceDowngrade(flags.ForceHTTP1()), http1DowngradeClient.ExtraH2ALPNs(alpn.PureGRPCALPNString), http1DowngradeClient.DialOpts(opts...))
 		}
+	} else if flags.ForceHTTP1() {
+		return nil, errors.New("cannot force HTTP/1 mode if direct gRPC is enabled")
 	}
 
 	apiToken, err := RetrieveAuthToken()